Update to latest drizzle version & separate relations from schema
