+++ package.json
@@ -54,8 +54,8 @@
 		"clsx": "^2.1.1",
 		"copy-to-clipboard": "^3.3.3",
 		"csstype": "^3.1.3",
-		"drizzle-orm": "^0.35.2",
-		"drizzle-zod": "^0.5.1",
+		"drizzle-orm": "^0.38.3",
+		"drizzle-zod": "^0.6.1",
 		"jotai": "^2.11.0",
 		"lodash.debounce": "^4.0.8",
 		"lucide-react": "^0.469.0",
@@ -86,7 +86,7 @@
 		"@types/react": "19.0.3",
 		"@types/react-dom": "19.0.2",
 		"autoprefixer": "^10.4.20",
-		"drizzle-kit": "^0.26.2",
+		"drizzle-kit": "^0.30.1",
 		"lefthook": "^1.10.1",
 		"postcss": "^8.4.49",
 		"serwist": "^9.0.11",
